src: resolve more conversion warnings on Windows
Some TBD POSSIBLE DATA LOSS scenarios marked.

+++ src/async-wrap.cc
@@ -208,7 +208,9 @@ void AsyncWrap::DestroyIdsCb(uv_idle_t* handle) {
     // Want each callback to be cleaned up after itself, instead of cleaning
     // them all up after the while() loop completes.
     HandleScope scope(env->isolate());
-    Local<Value> argv = Number::New(env->isolate(), current_id);
+    // TBD POSSIBLE DATA LOSS:
+    Local<Value> argv = Number::New(env->isolate(),
+                                    static_cast<double>(current_id));
     MaybeLocal<Value> ret = fn->Call(
         env->context(), Undefined(env->isolate()), 1, &argv);
 
@@ -259,14 +261,17 @@ AsyncWrap::AsyncWrap(Environment* env,
   HandleScope scope(env->isolate());
 
   Local<Value> argv[] = {
-    Number::New(env->isolate(), get_uid()),
+    // TBD POSSIBLE DATA LOSS:
+    Number::New(env->isolate(), static_cast<double>(get_uid())),
     Int32::New(env->isolate(), provider),
     Null(env->isolate()),
     Null(env->isolate())
   };
 
   if (parent != nullptr) {
-    argv[2] = Number::New(env->isolate(), parent->get_uid());
+    // TBD POSSIBLE DATA LOSS:
+    argv[2] = Number::New(env->isolate(),
+                          static_cast<double>(parent->get_uid()));
     argv[3] = parent->object();
   }
 
@@ -302,7 +307,9 @@ Local<Value> AsyncWrap::MakeCallback(const Local<Function> cb,
 
   Local<Function> pre_fn = env()->async_hooks_pre_function();
   Local<Function> post_fn = env()->async_hooks_post_function();
-  Local<Value> uid = Number::New(env()->isolate(), get_uid());
+  // TBD POSSIBLE DATA LOSS:
+  Local<Value> uid = Number::New(env()->isolate(),
+    static_cast<double>(get_uid()));
   Local<Object> context = object();
   Local<Object> domain;
   bool has_domain = false;
