+++ src/node_i18n.cc
@@ -115,7 +115,8 @@ struct Converter {
     conv = ucnv_open(name, &status);
     CHECK(U_SUCCESS(status));
     if (sub != NULL) {
-      ucnv_setSubstChars(conv, sub, strlen(sub), &status);
+      // TBD POSSIBLE DATA LOSS:
+      ucnv_setSubstChars(conv, sub, static_cast<int8_t>(strlen(sub)), &status);
     }
   }
